backend/services/export-service/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
sqlalchemy==2.0.10
asyncpg==0.27.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
pydantic==2.10.6
pandas==2.0.3
openpyxl==3.1.2
xlsxwriter==3.1.2
httpx==0.24.1
redis==4.5.3
celery==5.2.7 